feat: Removed identifiers for default vertices (#34)
* feat: Removed identifiers for default vertices

* feat: Separate passes for identity!

* feat: Instance rendering! (Partial support)

* feat: Small fixes and cleanups

* feat: Really small changes

Not really anything of value

+++ Engine/src/glfw/GLFWClock.cpp
@@ -4,7 +4,7 @@
 
 namespace Disarray {
 
-auto Clock::ms() -> float { return static_cast<float>(glfwGetTime()) * 1000.0F; }
-auto Clock::ns() -> float { return static_cast<float>(glfwGetTime()) * 1E6F; }
+auto Clock::ms() -> double { return glfwGetTime() * 1000.0F; }
+auto Clock::ns() -> double { return glfwGetTime() * 1E6F; }
 
 } // namespace Disarray
